requirements.txt
# Python >= 3.8

torch>=1.8.0
scikit-learn
xgboost
pandas
numpy
scipy
matplotlib
jupyterlab
PyYAML
joblib
tqdm
adversarial-robustness-toolbox